+++ tools/src/main/java/it/eng/tools/event/AuditEventType.java
@@ -12,6 +12,7 @@ public enum AuditEventType {
     APPLICATION_STOP("Application stop"),
     APPLICATION_LOGIN("Login"),
     APPLICATION_LOGOUT("Logout"),
+    PROTOCOL_NEGOTIATION_CONTRACT_NEGOTIATION("Contract negotiation"),
     PROTOCOL_NEGOTIATION_NOT_FOUND("Contract negotiation not found"),
     PROTOCOL_NEGOTIATION_STATE_TRANSITION_ERROR("State transition invalid"),
     PROTOCOL_NEGOTIATION_REQUESTED("Protocol negotiation requested"),